more setup for next days

+++ day09/src/part_two.rs
@@ -1,6 +1,6 @@
 use crate::problem::Problem;
 
-pub fn part_two(problem: &Problem) -> usize {
+pub fn part_two(problem: &Problem) -> u64 {
     // Implement the logic for part two here
     0
 }
