+++ day12/src/part_one.rs
@@ -1,6 +1,6 @@
 use crate::problem::Problem;
 
-pub fn part_one(problem: &Problem) -> usize {
+pub fn part_one(problem: &Problem) -> u64 {
     // Implement the logic for part one here
     0
 }
